removed programs and reverted to old structure
