+++ app/admin/components/layout/MobileNav.js
@@ -11,7 +11,6 @@ export default function MobileNav({ activeTab, setActiveTab, handleLogout }) {
     { id: "trainers", label: "Trainers" },
     { id: "schedules", label: "Schedule" },
     { id: "offeredClasses", label: "Classes" },
-    { id: "offeredPrograms", label: "Programs" },
     { id: "settings", label: "Settings" },
   ];
