+++ app/admin/page.js
@@ -11,7 +11,6 @@ import ConfirmModal from "./components/ui/ConfirmModal";
 import TrainersManager from "./components/trainers/TrainersManager";
 import ScheduleManager from "./components/schedules/ScheduleManager";
 import ClassManager from "./components/classes/ClassManager";
-import ProgramManager from "./components/programs/ProgramManager";
 import SettingsManager from "./components/SettingsManager";
 
 // Import custom hooks
@@ -19,7 +18,6 @@ import useAuth from "./hooks/useAuth";
 import useTrainers from "./hooks/useTrainers";
 import useSchedules from "./hooks/useSchedules";
 import useClasses from "./hooks/useClasses";
-import usePrograms from "./hooks/usePrograms";
 import useConfirmModal from "./hooks/useConfirmModal";
 
 
@@ -36,7 +34,7 @@ export default function AdminPage() {
     if (typeof window !== 'undefined') {
       const params = new URLSearchParams(window.location.search);
       const tabParam = params.get('tab');
-      if (tabParam && ['trainers', 'schedules', 'offeredClasses', 'offeredPrograms', 'settings'].includes(tabParam)) {
+      if (tabParam && ['trainers', 'schedules', 'offeredClasses', 'settings'].includes(tabParam)) {
         setActiveTab(tabParam);
       }
     }
@@ -72,15 +70,6 @@ export default function AdminPage() {
   } = useClasses();
 
   const {
-    programs: offeredPrograms,
-    isLoading: isProgramsLoading,
-    isSubmitting: isAddingProgram,
-    isDeleting: isDeletingProgram,
-    addProgram,
-    deleteProgram
-  } = usePrograms();
-
-  const {
     confirmModal,
     openConfirmModal,
     closeConfirmModal,
@@ -98,9 +87,6 @@ export default function AdminPage() {
     offeredClasses: isClassesLoading,
     addOfferedClass: isAddingClass,
     deleteOfferedClass: isDeletingClass,
-    offeredPrograms: isProgramsLoading,
-    addOfferedProgram: isAddingProgram,
-    deleteOfferedProgram: isDeletingProgram
   };
 
   // Handler functions for confirmation modals
@@ -127,15 +113,6 @@ export default function AdminPage() {
       title: "Delete Class",
       message: "Are you sure you want to delete this class? This action cannot be undone.",
       onConfirm: deleteClass,
-      itemId: id
-    });
-  };
-
-  const confirmDeleteProgram = (id) => {
-    openConfirmModal({
-      title: "Delete Program",
-      message: "Are you sure you want to delete this program? This action cannot be undone.",
-      onConfirm: deleteProgram,
       itemId: id
     });
   };
@@ -242,15 +219,6 @@ export default function AdminPage() {
                   isLoading={isLoading}
                   handleAddClass={addClass}
                   confirmDeleteClass={confirmDeleteClass}
-                />
-              )}
-
-              {activeTab === "offeredPrograms" && (
-                <ProgramManager
-                  programs={offeredPrograms}
-                  isLoading={isLoading}
-                  handleAddProgram={addProgram}
-                  confirmDeleteProgram={confirmDeleteProgram}
                 />
               )}
