--- components/OfferingsSlider.js
@@ -1,355 +1,0 @@
-"use client";
-import { useState, useRef, useEffect, useCallback } from "react";
-import { motion, AnimatePresence } from "framer-motion";
-import Image from "next/image";
-import dynamic from "next/dynamic";
-
-const ClassDetailsModal = dynamic(() => import("./ClassDetailsModal"), { ssr: false });
-const ProgramDetailsModal = dynamic(() => import("./ProgramDetailsModal"), { ssr: false });
-
-export default function OfferingsSlider({ initialClasses, initialPrograms, isDark }) {
-  const [activeTab, setActiveTab] = useState("classes");
-  const [classes, setClasses] = useState(initialClasses || []);
-  const [programs, setPrograms] = useState(initialPrograms || []);
-  const [loading, setLoading] = useState(!initialClasses && !initialPrograms);
-  const [width, setWidth] = useState(0);
-  const [isDragging, setIsDragging] = useState(false);
-  const [selectedClass, setSelectedClass] = useState(null);
-  const [selectedProgram, setSelectedProgram] = useState(null);
-  const [isClassModalOpen, setIsClassModalOpen] = useState(false);
-  const [isProgramModalOpen, setIsProgramModalOpen] = useState(false);
-  const sliderRef = useRef(null);
-  const contentRef = useRef(null);
-  const dragStartTime = useRef(0);
-  const dragDistance = useRef(0);
-  const dragStartPos = useRef(0);
-
-  const currentItems = activeTab === "classes" ? classes : programs;
-
-  // Callback ref: recalculate drag width when new content DOM mounts
-  const handleContentRef = useCallback((node) => {
-    contentRef.current = node;
-    if (node) {
-      requestAnimationFrame(() => {
-        if (sliderRef.current && contentRef.current) {
-          const newWidth = contentRef.current.scrollWidth - sliderRef.current.offsetWidth;
-          setWidth(Math.max(0, newWidth));
-        }
-      });
-    }
-  }, []);
-
-  // Fetch data only if not provided via props
-  useEffect(() => {
-    if (initialClasses && initialPrograms) return;
-
-    const fetchData = async () => {
-      try {
-        setLoading(true);
-        const [classesRes, programsRes] = await Promise.all([
-          fetch("/api/classes"),
-          fetch("/api/programs"),
-        ]);
-
-        const classesData = await classesRes.json();
-        const programsData = await programsRes.json();
-
-        if (classesData.success) {
-          setClasses(classesData.classes || []);
-        }
-        if (programsData.success) {
-          const activePrograms = (programsData.programs || []).filter((prog) => prog.active);
-          setPrograms(activePrograms);
-        }
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, [initialClasses, initialPrograms]);
-
-  // Calculate width
-  useEffect(() => {
-    if (sliderRef.current && contentRef.current && currentItems?.length) {
-      const newWidth = contentRef.current.scrollWidth - sliderRef.current.offsetWidth;
-      setWidth(Math.max(0, newWidth));
-    }
-  }, [currentItems, activeTab]);
-
-  // Update width on resize
-  useEffect(() => {
-    const handleResize = () => {
-      if (sliderRef.current && contentRef.current && currentItems?.length) {
-        const newWidth = contentRef.current.scrollWidth - sliderRef.current.offsetWidth;
-        setWidth(Math.max(0, newWidth));
-      }
-    };
-
-    window.addEventListener("resize", handleResize);
-    return () => window.removeEventListener("resize", handleResize);
-  }, [currentItems]);
-
-  const handleDragStart = useCallback((e, info) => {
-    setIsDragging(true);
-    dragStartTime.current = Date.now();
-    dragStartPos.current = info.point.x;
-    dragDistance.current = 0;
-
-    document.querySelectorAll(".offering-card-link").forEach((link) => {
-      link.style.pointerEvents = "none";
-    });
-  }, []);
-
-  const handleDragEnd = useCallback((e, info) => {
-    dragDistance.current = Math.abs(info.point.x - dragStartPos.current);
-
-    setTimeout(() => {
-      setIsDragging(false);
-      document.querySelectorAll(".offering-card-link").forEach((link) => {
-        link.style.pointerEvents = "auto";
-      });
-    }, 50);
-  }, []);
-
-  const handleMouseDown = useCallback((e) => {
-    if (e.button === 0) {
-      e.preventDefault();
-    }
-  }, []);
-
-  const handleCardClick = useCallback(
-    (e, item) => {
-      const dragTime = Date.now() - dragStartTime.current;
-      if (!isDragging || (dragDistance.current < 20 && dragTime < 200)) {
-        e.preventDefault();
-        if (activeTab === "classes") {
-          setSelectedClass(item);
-          setIsClassModalOpen(true);
-        } else {
-          setSelectedProgram(item);
-          setIsProgramModalOpen(true);
-        }
-      }
-    },
-    [isDragging, activeTab]
-  );
-
-  const handleImageError = useCallback((e) => {
-    e.target.onerror = null;
-    e.target.style.backgroundColor = "#1A1A1A";
-  }, []);
-
-  if (loading) {
-    return (
-      <div className="max-w-screen-2xl mx-auto px-4 py-12">
-        <div className="flex justify-center py-12">
-          <div className="h-12 w-12 border-t-2 border-b-2 border-[#E50914] animate-spin" />
-        </div>
-      </div>
-    );
-  }
-
-  const hasClasses = classes.length > 0;
-  const hasPrograms = programs.length > 0;
-
-  // If only one type exists, don't show tabs
-  if (!hasClasses && !hasPrograms) {
-    return (
-      <div className="max-w-screen-2xl mx-auto px-4 py-12 text-center">
-        <p className="text-[rgba(15,15,15,0.55)] dark:text-[rgba(237,235,230,0.6)]">
-          No offerings available at the moment.
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <div className="max-w-screen-2xl mx-auto px-0 py-12">
-      {/* Tab Switcher — brutalist bar, sharp corners */}
-      {hasClasses && hasPrograms && (
-        <div className="flex justify-center mb-10">
-          <div className="inline-flex bg-[#F0F0F0] dark:bg-[#1A1A1A] p-1">
-            <button
-              onClick={() => setActiveTab("classes")}
-              className={`relative px-8 py-3 text-lg font-bold transition-all duration-300 ${
-                activeTab === "classes"
-                  ? "text-white"
-                  : "text-[rgba(15,15,15,0.55)] dark:text-[rgba(237,235,230,0.6)] hover:text-[#1A1A1A] dark:hover:text-[#EDEBE6]"
-              }`}
-            >
-              {activeTab === "classes" && (
-                <motion.div
-                  layoutId="activeTab"
-                  className="absolute inset-0 bg-[#E50914]"
-                  transition={{ type: "spring", bounce: 0.2, duration: 0.6 }}
-                />
-              )}
-              <span className="relative z-10 uppercase tracking-wide">Classes</span>
-            </button>
-            <button
-              onClick={() => setActiveTab("programs")}
-              className={`relative px-8 py-3 text-lg font-bold transition-all duration-300 ${
-                activeTab === "programs"
-                  ? "text-white"
-                  : "text-[rgba(15,15,15,0.55)] dark:text-[rgba(237,235,230,0.6)] hover:text-[#1A1A1A] dark:hover:text-[#EDEBE6]"
-              }`}
-            >
-              {activeTab === "programs" && (
-                <motion.div
-                  layoutId="activeTab"
-                  className="absolute inset-0 bg-[#E50914]"
-                  transition={{ type: "spring", bounce: 0.2, duration: 0.6 }}
-                />
-              )}
-              <span className="relative z-10 uppercase tracking-wide">Programs</span>
-            </button>
-          </div>
-        </div>
-      )}
-
-      {/* Section Title */}
-      <motion.h2
-        key={activeTab}
-        initial={{ opacity: 0, y: -20 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ duration: 0.4 }}
-        className="font-display uppercase text-center mb-12 text-[#1A1A1A] dark:text-[#EDEBE6]"
-        style={{
-          fontSize: "clamp(2.5rem, 5vw, 4.5rem)",
-          lineHeight: 0.88,
-          letterSpacing: "-0.04em",
-        }}
-      >
-        {activeTab === "classes" ? "OUR CLASSES" : "OUR PROGRAMS"}
-      </motion.h2>
-
-      {/* Slider */}
-      <AnimatePresence mode="wait">
-        <motion.div
-          key={activeTab}
-          initial={{ opacity: 0, x: 20 }}
-          animate={{ opacity: 1, x: 0 }}
-          exit={{ opacity: 0, x: -20 }}
-          transition={{ duration: 0.3 }}
-          className="overflow-hidden relative"
-        >
-          {currentItems.length === 0 ? (
-            <div className="text-center py-8 text-[rgba(15,15,15,0.55)] dark:text-[rgba(237,235,230,0.6)]">
-              No {activeTab} available at the moment.
-            </div>
-          ) : (
-            <>
-              <motion.div
-                ref={sliderRef}
-                className="cursor-grab overflow-hidden"
-                whileTap={{ cursor: "grabbing" }}
-                onMouseDown={handleMouseDown}
-              >
-                <motion.div
-                  ref={handleContentRef}
-                  className="flex"
-                  drag="x"
-                  dragConstraints={{ right: 0, left: -width }}
-                  dragTransition={{
-                    power: 0.25,
-                    timeConstant: 400,
-                    modifyTarget: (target) => Math.round(target / 100) * 100,
-                  }}
-                  onDragStart={handleDragStart}
-                  onDragEnd={handleDragEnd}
-                  dragElastic={0.05}
-                  dragMomentum={true}
-                >
-                  {currentItems.map((item) => (
-                    <motion.div
-                      key={item._id}
-                      className="min-w-[270px] sm:min-w-[350px] md:min-w-[400px] lg:min-w-[420px] h-[500px] relative px-2"
-                      whileHover={{ scale: isDragging ? 1 : 1.02 }}
-                      transition={{ duration: 0.3 }}
-                    >
-                      <div
-                        className="offering-card-link block h-full w-full relative cursor-pointer overflow-hidden"
-                        onClick={(e) => handleCardClick(e, item)}
-                      >
-                        {/* Bottom gradient overlay — sharp, to dark surface */}
-                        <div
-                          className="absolute inset-0 z-10"
-                          style={{
-                            background:
-                              "linear-gradient(to top, #141414 0%, rgba(20,20,20,0.6) 40%, transparent 100%)",
-                          }}
-                        />
-                        <div className="relative h-full w-full">
-                          <Image
-                            src={item.image || "/images/default-class.jpg"}
-                            alt={item.title}
-                            fill
-                            className="object-cover w-full h-full transition-transform duration-700 group-hover:scale-110"
-                            sizes="(max-width: 768px) 90vw, (max-width: 1200px) 40vw, 33vw"
-                            onError={handleImageError}
-                            draggable="false"
-                          />
-                        </div>
-                        <div className="absolute bottom-0 left-0 right-0 z-20 p-6 flex flex-col items-center">
-                          <h3
-                            className="font-display uppercase text-white text-center mb-2"
-                            style={{
-                              fontSize: "clamp(1.5rem, 3vw, 2.5rem)",
-                              lineHeight: 0.88,
-                              letterSpacing: "-0.04em",
-                            }}
-                          >
-                            {item.title}
-                          </h3>
-                          {item.shortDescription && (
-                            <p className="text-[rgba(237,235,230,0.6)] text-base mt-1 mb-3 text-center line-clamp-2">
-                              {item.shortDescription}
-                            </p>
-                          )}
-                        </div>
-                      </div>
-                    </motion.div>
-                  ))}
-                </motion.div>
-              </motion.div>
-
-              {/* Scroll hint — minimal brutalist dots */}
-              {currentItems.length > 2 && (
-                <>
-                  <div className="absolute top-1/2 right-3 -translate-y-1/2 flex flex-col gap-1.5 opacity-70 z-20">
-                    <div className="w-1.5 h-1.5 bg-white" />
-                    <div className="w-1.5 h-1.5 bg-white" />
-                    <div className="w-1.5 h-1.5 bg-white" />
-                  </div>
-                  <motion.p
-                    initial={{ opacity: 0 }}
-                    animate={{ opacity: 0.7 }}
-                    transition={{ delay: 0.5, duration: 0.6 }}
-                    className="text-center text-sm uppercase tracking-wide mt-4 text-[rgba(15,15,15,0.55)] dark:text-[rgba(237,235,230,0.6)]"
-                  >
-                    Drag to explore more {activeTab}
-                  </motion.p>
-                </>
-              )}
-            </>
-          )}
-        </motion.div>
-      </AnimatePresence>
-
-      {/* Modals */}
-      <ClassDetailsModal
-        isOpen={isClassModalOpen}
-        onClose={() => setIsClassModalOpen(false)}
-        classData={selectedClass}
-      />
-      <ProgramDetailsModal
-        isOpen={isProgramModalOpen}
-        onClose={() => setIsProgramModalOpen(false)}
-        programData={selectedProgram}
-      />
-    </div>
-  );
-}
