+++ lib/Class.js
@@ -22,22 +22,22 @@ const ClassSchema = new mongoose.Schema({
     maxlength: [200, 'Short description cannot exceed 200 characters']
   },
   pricing: {
-    walkIn: {
+    regular: {
       type: Number,
       default: 0,
       min: [0, 'Price cannot be negative']
     },
-    weekly: {
+    student: {
       type: Number,
       default: 0,
       min: [0, 'Price cannot be negative']
     },
-    monthly: {
+    kids: {
       type: Number,
       default: 0,
       min: [0, 'Price cannot be negative']
     },
-    annual: {
+    women: {
       type: Number,
       default: 0,
       min: [0, 'Price cannot be negative']
