+++ lib/data.js
@@ -2,7 +2,6 @@ import { cache } from 'react';
 import { connectDB } from './dbConnect';
 import Trainer from './Trainer';
 import Class from './Class';
-import Program from './Program';
 import Schedule from './Schedule';
 import Settings from './Settings';
 
@@ -16,10 +15,10 @@ export const getClasses = cache(async () => {
   return serialize(classes);
 });
 
-export const getActivePrograms = cache(async () => {
+export const getActiveClasses = cache(async () => {
   await connectDB();
-  const programs = await Program.find({ active: true }).lean();
-  return serialize(programs);
+  const classes = await Class.find({ active: true }).lean();
+  return serialize(classes);
 });
 
 export const getTrainers = cache(async () => {
